Basic User paging
- some slight changes to event with work to take in Template t instead of hard coding to int
- User Virtual memory page table creation
- mapping vaddr to a page with map_vaddr
- use page table with pagetable::use_page_table

+++ include/frame.h
@@ -11,7 +11,7 @@
 
 void create_frame_table(uintptr_t start, int size);
 
-void alloc_frame(int flags, Function<void(int)> w);
+void alloc_frame(int flags, Function<void(uint64_t)> w);
 
 bool free_frame(uintptr_t frame_addr);
